fix(binder): distinguish instantiated vs uninstantiated namespace modules
`modules/binding.rs` set `VALUE_MODULE | NAMESPACE_MODULE` unconditionally on
every identifier-named namespace, so nothing downstream could tell an
instantiated module (`namespace X { var v = 1 }`) from an uninstantiated one
(`namespace X {}` or type-only). Wire the existing syntactic
`NodeArena::is_namespace_instantiated` classifier (already used by the
checker, implements tsc's `getModuleInstanceState`) into the binder's own
flag assignment: identifier-named namespaces now get `VALUE_MODULE` xor
`NAMESPACE_MODULE`; ambient string-literal modules are unchanged (tsc always
treats those as instantiated regardless of body).

This exposed a latent gap in `can_merge_flags`: it allowed TYPE_ALIAS to
merge with a namespace only via the VALUE-based rule, which accidentally
worked before because VALUE_MODULE was always set. Added an explicit
TYPE_ALIAS+MODULE merge rule so `namespace Pick {} type Pick<T,K> = T;`
(oracled against tsc, legal per `TypeAliasExcludes`) keeps merging correctly
now that an uninstantiated namespace no longer carries VALUE_MODULE.

Also switches `merge_namespace_exports_into_function`'s `typeof f` gate
(#16133) from a member-count heuristic to reading the merged symbol's
VALUE_MODULE flag directly, and un-ignores the test #16133 pinned against
this exact gap (`ts2322_namespace_with_only_unexported_value_still_renders_typeof`).

Closes #16136.
