+++ crates/tsz-binder/src/nodes/mod.rs
@@ -8,4 +8,5 @@ mod binding;
 mod binding_scope;
 mod flow_statements;
 mod hoisting;
+mod merge_flags;
 mod names;
